update system e-stop
