+++ panther_hardware_interfaces/src/panther_system/system_e_stop.cpp
@@ -17,12 +17,12 @@
 #include <memory>
 #include <string>
 
-#include "panther_hardware_interfaces/panther_system/panther_system_e_stop.hpp"
+#include "panther_hardware_interfaces/panther_system/system_e_stop.hpp"
 
 namespace panther_hardware_interfaces
 {
 
-bool EStopPTH12X::ReadEStopState()
+bool EStop::ReadEStopState()
 {
   if (e_stop_manipulation_mtx_.try_lock()) {
     std::lock_guard<std::mutex> e_stop_lck(e_stop_manipulation_mtx_, std::adopt_lock);
@@ -39,7 +39,7 @@ bool EStopPTH12X::ReadEStopState()
   return e_stop_triggered_;
 }
 
-void EStopPTH12X::TriggerEStop()
+void EStop::TriggerEStop()
 {
   gpio_controller_->InterruptEStopReset();
 
@@ -53,7 +53,7 @@ void EStopPTH12X::TriggerEStop()
   e_stop_triggered_ = true;
 }
 
-void EStopPTH12X::ResetEStop()
+void EStop::ResetEStop()
 {
   if (e_stop_manipulation_mtx_.try_lock()) {
     std::lock_guard<std::mutex> e_stop_lck(e_stop_manipulation_mtx_, std::adopt_lock);
@@ -73,63 +73,6 @@ void EStopPTH12X::ResetEStop()
     } catch (const std::runtime_error & e) {
       throw std::runtime_error(
         "Error when trying to reset E-stop using GPIO: " + std::string(e.what()));
-    }
-
-    roboteq_error_filter_->SetClearErrorsFlag();
-
-    e_stop_triggered_ = false;
-  } else {
-    std::fprintf(stdout, "E-stop trigger operation is in progress, skipping reset.");
-  }
-}
-
-bool EStopPTH10X::ReadEStopState()
-{
-  if (e_stop_manipulation_mtx_.try_lock()) {
-    std::lock_guard<std::mutex> e_stop_lck(e_stop_manipulation_mtx_, std::adopt_lock);
-    const bool motors_on = gpio_controller_->IsPinActive(GPIOPin::STAGE2_INPUT);
-    const bool driver_error = roboteq_error_filter_->IsError();
-
-    if ((driver_error || !motors_on) && !e_stop_triggered_) {
-      TriggerEStop();
-    }
-  }
-
-  return e_stop_triggered_;
-}
-
-void EStopPTH10X::TriggerEStop()
-{
-  std::lock_guard<std::mutex> e_stop_lck(e_stop_manipulation_mtx_);
-  std::lock_guard<std::mutex> lck_g(*motor_controller_write_mtx_);
-
-  try {
-    motors_controller_->TurnOnSafetyStop();
-  } catch (const std::runtime_error & e) {
-    throw std::runtime_error(
-      "Error when trying to set safety stop using CAN command: " + std::string(e.what()));
-  }
-
-  e_stop_triggered_ = true;
-}
-
-void EStopPTH10X::ResetEStop()
-{
-  if (e_stop_manipulation_mtx_.try_lock()) {
-    std::lock_guard<std::mutex> e_stop_lck(e_stop_manipulation_mtx_, std::adopt_lock);
-
-    if (!ZeroVelocityCheck()) {
-      throw std::runtime_error(
-        "Can't reset E-stop - last velocity commands are different than zero. Make sure that your "
-        "controller sends zero commands before trying to reset E-stop.");
-    }
-
-    if (!gpio_controller_->IsPinActive(GPIOPin::STAGE2_INPUT)) {
-      throw std::runtime_error("Can't reset E-stop - motors are not powered.");
-    }
-
-    if (roboteq_error_filter_->IsError()) {
-      throw std::runtime_error("Can't reset E-stop - motor controller is in an error state.");
     }
 
     roboteq_error_filter_->SetClearErrorsFlag();
